Generic BST Node and Generic BST
Functional generic BST node and tree classes, I still need to comment
everything but it works so far, still need to add methods. This ones
should work with anything other than Integer but for objects you need to
override compareTo. As for Integers it works perfectly so far at least
for what we need.